Suggested changes for README.MD
Tensorflow 1.9 won't install for python 3.8. It requires python 3.6.

It's unecessery to install scipy and numpy because it's referenced by other libraries so the versions will be downgraded (numpy to 1.14.2 and scipy to 1.1.0).

Couldn't instlal gensim by conda (many versioning errors) so i installed it's latest version by pip and than lesson 5 starts to work perfectly.